requirements.txt
django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2